requirements.txt
atomic-store==0.0.1
pillow==11.2.1
python-telegram-bot[job-queue]==22.1
